renamed util.js class to Frame and renamed the exporting class
