+++ samples/charts/Frame.js
@@ -26,7 +26,7 @@ import getMuiTheme from 'material-ui/styles/getMuiTheme';
 import { Card, CardActions, CardHeader, CardMedia, CardTitle, CardText } from 'material-ui/Card';
 import FlatButton from 'material-ui/FlatButton';
 
-export class Row extends React.Component {
+export class Frame extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
