+++ samples/charts/ScatterPlotConfigSample.js
@@ -17,7 +17,7 @@
  */
 
 import React from 'react';
-import { Row } from './util';
+import { Row } from './Frame';
 import './css/Table.css';
 import VizG from '../../src/VizG.jsx';
